refactor(core): merge modules
