+++ crates/core/src/agent/event.rs
