+++ crates/core/src/agent/tool.rs
