+++ crates/core/src/runtime/mod.rs
@@ -1,0 +1,3 @@
+//! Runtime primitives for agent execution and tool dispatch.
+
+pub mod hook;
